history payment

+++ ASM.Services/Models/Mepa/PaymentLinkResponse.cs
@@ -5,7 +5,7 @@ namespace ASM.Services.Models.Mepa
     public class PaymentLinkResponse : RequestResponseBase
     {
         public string id { get; set; }
-        public double? price { get; set; }
+        public decimal? price { get; set; }
         public string? init_point { get; set; }
     }
 }
